Switching to raw strings (https://doc.rust-lang.org/reference/tokens.html#raw-string-literals); all regex examples use them (https://doc.rust-lang.org/regex/regex/index.html)
Without raw strings, to recognize a left parenthesis in a regex, we would need
to use the string \\( in the rust program.  The first \ is used to escape \,
which passes the string \( to the regex. The second \ is required to escape
the ( because ( has a special meaning within Rust regexes. With raw strings, we
can say \( instead.