Fairer benchmark; ~0.07 seconds for iterator
This (https://doc.rust-lang.org/src/alloc/vec.rs.html#767-786) explains why remove is so expensive. It's because it involves (1) shifting the contents of the vector using ptr::copy, which is equivalent to C's memmove, and (2) updating the vector length. As an aside, it doesn't look like vector.remove() actually frees up the memory though.